+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - add more information to errors
   - Add unique ID checks
   - Add path to output
+- GeoJSON to JSON:
+  - Include name in organisation and phase references
 
 ## [0.4.0] - 2022-11-09